+++ Tests/BFKitTests/Linux/Foundation/NSPointerArrayExtensionTests.swift
@@ -1,0 +1,112 @@
+//
+//  NSPointerArrayExtensionTests.swift
+//  BFKit
+//
+//  The MIT License (MIT)
+//
+//  Copyright (c) 2015 - 2017 Fabrizio Brancati. All rights reserved.
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in all
+//  copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+//  SOFTWARE.
+
+import XCTest
+import Foundation
+@testable import BFKit
+
+class NSPointerArrayExtensionTests: XCTestCase {
+    static let allTests = [
+        ("testAdd", testAdd),
+        ("testInsert", testInsert),
+        ("testReplace", testRepace),
+        ("testGet", testGet),
+        ("testRemove", testRemove)
+    ]
+    
+    let array = NSPointerArray.weakObjects()
+    
+    override func setUp() {
+        super.setUp()
+    }
+    
+    override func tearDown() {
+        super.tearDown()
+    }
+    
+    func testAdd() {
+        let array1: NSArray = NSArray(array: [1, 2])
+        let array2: NSArray = NSArray(array: [3, 4])
+        
+        array.add(array1)
+        array.add(array2)
+        
+        let array1Test: NSArray = array.object(at: 0) as! NSArray // swiftlint:disable:this force_cast
+        let array2Test: NSArray = array.object(at: 1) as! NSArray // swiftlint:disable:this force_cast
+        
+        XCTAssertEqual(array1Test, array1)
+        XCTAssertEqual(array2Test, array2)
+    }
+    
+    func testInsert() {
+        let array1: NSArray = NSArray(array: [1, 2])
+        let array2: NSArray = NSArray(array: [3, 4])
+        
+        array.add(array1)
+        array.insert(array2, at: 0)
+        
+        let array1Test: NSArray = array.object(at: 0) as! NSArray // swiftlint:disable:this force_cast
+        let array2Test: NSArray = array.object(at: 1) as! NSArray // swiftlint:disable:this force_cast
+        
+        XCTAssertEqual(array1Test, array2)
+        XCTAssertEqual(array2Test, array1)
+    }
+    
+    func testReplace() {
+        let array1: NSArray = NSArray(array: [1, 2])
+        let array2: NSArray = NSArray(array: [3, 4])
+        
+        array.add(array1)
+        array.replace(at: 0, with: array2)
+        
+        let array1Test: NSArray = array.object(at: 0) as! NSArray // swiftlint:disable:this force_cast
+        
+        XCTAssertEqual(array1Test, array2)
+    }
+    
+    func testGet() {
+        let array1: NSArray = NSArray(array: [1, 2])
+        
+        array.add(array1)
+        
+        let array1Test: NSArray = array.object(at: 0) as! NSArray // swiftlint:disable:this force_cast
+        
+        XCTAssertEqual(array1Test, array1)
+    }
+    
+    func testRemove() {
+        let array1: NSArray = NSArray(array: [1, 2])
+        let array2: NSArray = NSArray(array: [3, 4])
+        
+        array.add(array1)
+        array.add(array2)
+        array.remove(at: 0)
+        
+        let array2Test: NSArray = array.object(at: 0) as! NSArray // swiftlint:disable:this force_cast
+        
+        XCTAssertEqual(array2Test, array2)
+    }
+}