+++ CHANGELOG.md
@@ -31,6 +31,9 @@ All notable changes to this project will be documented in this file.<br>
 ## watchOS Support
 ### Added
 - Added watchOS support (For specific classes, structs and extensions support refer to [compatibility](https://github.com/FabrizioBrancati/BFKit-Swift/tree/master#compatibility) table)
+- Added `WKInterfaceController` class extension for watchOS
+- Added `NSPointerArray` class extension
+- Added `isNotEmpty` in Array extension
 - Added `saveLog(in:, filename:)` in BFLog struct to save log in a file
 
 ### Improved